supermarkt added filialenname minor fixes in Ware toString(), Constructor
