+++ de/ostfalia/practice/id568441/Kleidung.java
@@ -1,0 +1,9 @@
+package de.ostfalia.practice.id568441;
+
+public class Kleidung extends Ware {
+    private enum groesse {XS,S,M,L,XL,XXL}
+
+    public Kleidung(String filialenname, double preis, int artikelnummer, String name, short altersfreigabe, String hersteller) {
+        super(filialenname, preis, artikelnummer, name, altersfreigabe, hersteller);
+    }
+}
