+++ de/ostfalia/practice/id568441/supermarkt.java
@@ -1,0 +1,19 @@
+package de.ostfalia.practice.id568441;
+
+public class supermarkt {
+    private String filialenname;
+
+    public supermarkt(String filialenname) {
+        this.filialenname = filialenname;
+    }
+
+    public String getFilialenname() {
+        return filialenname;
+    }
+
+    public void setFilialenname(String filialenname) {
+        this.filialenname = filialenname;
+    }
+
+
+}
