--- supermarkt.java
@@ -1,5 +1,0 @@
-package de.ostfalia.practice.id568441;
-
-public class supermarkt {
-
-}
